Move open rosa implementation of FormSource
