+++ collect_app/src/main/java/org/odk/collect/android/formmanagement/FormSourceProvider.kt
@@ -1,8 +1,8 @@
 package org.odk.collect.android.formmanagement
 
-import org.odk.collect.android.openrosa.OpenRosaFormSource
 import org.odk.collect.android.utilities.WebCredentialsUtils
 import org.odk.collect.forms.FormSource
+import org.odk.collect.openrosa.forms.OpenRosaFormSource
 import org.odk.collect.openrosa.http.OpenRosaHttpInterface
 import org.odk.collect.openrosa.parse.OpenRosaResponseParserImpl
 import org.odk.collect.projects.ProjectDependencyFactory
