+++ open-rosa/src/test/java/org/odk/collect/openrosa/http/OpenRosaHeadRequestTest.java
@@ -10,7 +10,7 @@ import static org.junit.Assert.fail;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-import org.odk.collect.openrosa.http.support.MockWebServerRule;
+import org.odk.collect.openrosa.support.MockWebServerRule;
 
 import java.net.URI;
