+++ open-rosa/src/test/java/org/odk/collect/openrosa/support/MockWebServerRule.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.openrosa.http.support
+package org.odk.collect.openrosa.support
 
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.rules.TestRule
